requirements.txt
altgraph==0.17.4
asgiref==3.8.1
bcc==0.18.0
blinker==1.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
docopt==0.6.2
evdev==1.6.1
filelock==3.13.1
Flask==3.0.2
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
gyp==0.1
hidpidaemon==18.4.6
httplib2==0.20.2
icalendar==5.0.11
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.6
MarkupSafe==2.1.5
maskpass==0.3.7
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==23.2
platformdirs==4.2.0
pop-transition==1.1.2
prettytable==3.9.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
pyfiglet==1.0.2
PyGObject==3.42.1
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sqlparse==0.5.0
systemd-python==234
tabulate==0.9.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
